refactoring, use c_map module name
