+++ src/redisearch_rs/trie_bencher/src/lib.rs
@@ -7,13 +7,13 @@ use std::{
 pub use bencher::OperationBencher;
 
 pub mod bencher;
-pub mod cmap;
+pub mod c_map;
 pub mod corpus;
 pub mod ffi;
 mod redis_allocator;
 
 // Convenient aliases for the trie types that are being benchmarked.
-pub use cmap::CTrieMap;
+pub use c_map::CTrieMap;
 pub type RustTrieMap = trie_rs::trie::TrieMap<NonNull<c_void>>;
 
 /// Convert a string to a slice of `c_char`, allocated on the heap, which is the expected input for [crate::RustTrieMap].
